Move data generated by `MoveTestDataSet` to VTK data files
The `MoveTestDataSet` predates the ability to load VTK data files as
well as the testing conventions for loading data files from the
`vtk-m/data` directory. This method of creating data is a bit silly now,
so instead add small data files to the VTK-m repo that can be loaded by
the testing framework.