Remove Invocation from WorkletInvokeFunctor
We are on a quest to remove the `Invocation` from the dispatcher's
invoke because it requires an unnecessary amount of template
information. The first step is to remake WorkletInvokeFunctor to not
require it. (It is now called directly with the parameter arguments and
the worklet to call.) This implementation removes a lot of generated
code.