Use named functor classes for Load/Store of worklet arguments
Previously, anonymous lambda functions were used to load and store each
worklet argument wiht the associated fetch. This was fine with C++14
syntax and was working in most places. However, the MSVC compiler was
crashing with a SEGFAULT with no indication of where or why. I suspect
there is some sort of compiler bug here, perhaps with the captures for
the lambda function. To try to get around this, convert the lambda
functions to named structs and hope that works around any compiler
idiosyncrasies.