Remove Invoke from DispatcherBase::BasicInvoke
The `BasicInvoke` method previously used the `Invoke` class to transport
arguments to the execution environment and then schedule the work on the
device. This has been changed to pass the `Invoke` arguments directly
and use variadic template tricks to transport and schedule. (The old
interface is still there, but will be deleted soon.)

Part of this work includes removing the old forms of MakeTask from the
device implementations.

It should be noted that because objects are not taken from the
`Invocation` object, they seem to be coming out as const references.
Several `Transport` and `ExecutionObjectBase` classes had to be changed
to accomodate `const` control objects.