Remove return types for auto methods in Tuple
In C++11, if you had a method that returned `auto`, you had to declare
what the type it actually returned was using a `->` after the function
prototype. This is no longer required in C++14, and the compiler can
automatically determine the return type. As such, we can clean up the
code by removing these unnecessary and messy return types.

For a practical reason, declaring the return type can make for some odd
SFINAE effects that are hard to debug. In fact, I was getting some weird
behavior where the return type declaration was causing the `const` form
of `Apply` to be called instead of the expected non-const version.

I've also added a bunch of `VTKM_NEVER_EXPORT` modifiers to the `Tuple`
classes, methods, and functions. Typically, there is no reason to export
these templated symbols from anything, and doing so will simply require
larger symbol tables. I don't know if it will have any practical effect,
but it should not hurt.