Support lambda functions in ArrayHandleImplicit
`ArrayHandleImplicit` now supports using a lambda function as its functor.
Using lambdas, implicit arrays are easily constructed with
`make_ArrayHandleImplicit`.

``` cpp
auto implicitArray = make_ArrayHandleImplicit(
  VTKM_LAMBDA(vtkm::Id index) { return /* value based on index */; }, size);
```

Previously, lambdas did not work with `ArrayHandleImplicit because the
object created for a lambda function does not have a default constructor.
The need for a default constructor has been eliminated.

It should also be noted that using a lambda in `ArrayHandleImplicit`
means that it has to be compiled for both control and execution
environments. To allow that to happen, a `VTKM_LAMBDA` macros was added.