Use ArrayHandles for flying edges lookup tables
The previous implementation declared a static array and looked up
values in functions. This leaves the compiler to manage to data.
That can have potential advantages such as putting data in
faster constant data. But it could also cause problems if the
compiler has problems managing all the data.

The switch to ArrayHandles simplifies the implementation for
the compiler by managing these static arrays on the host and
using the existing array management to make sure everything
is correct.